Update How To Wabbajack.md
Expanded the Preinstallation steps to include creating folder.

Lots of general edits to improve readability, correct grammar or spelling errors, and provide expanded details.

removed the launching the game section.  This will be moved to the main readme.  Users should complete the post install steps before launching.

removed the FO4LODGen reference as that is not used by this list.

Updated the installation problems section.